Renamed types->utils
